Working merge now. Ready for refactor into a channel to drop mutex
