+++ src/lib.rs
@@ -1,4 +1,3 @@
-#![feature(integer_atomics)]
 extern crate portaudio;
 pub mod analyze;
 pub mod input_output_setup;
@@ -8,3 +7,4 @@ pub mod ring_buffer;
 pub mod settings;
 pub mod sine;
 pub mod state;
+pub mod ratios;
